Move certificate_aims autocomplete to education_group app
