+++ education_group/tests/views/autocomplete.py
@@ -29,9 +29,11 @@ from django.test import TestCase
 from django.urls import reverse
 
 from base.models.enums.education_group_categories import TRAINING, MINI_TRAINING, GROUP
+from base.tests.factories.certificate_aim import CertificateAimFactory
 from base.tests.factories.education_group_type import EducationGroupTypeFactory, MiniTrainingEducationGroupTypeFactory,\
     GroupEducationGroupTypeFactory
 from base.tests.factories.person import PersonFactory
+from base.tests.factories.user import SuperUserFactory
 
 
 class TestEducationGroupTypeAutoComplete(TestCase):
@@ -74,3 +76,46 @@ class TestEducationGroupTypeAutoComplete(TestCase):
         }
         self.assertEqual(len(json_response["results"]), 1)
         self.assertEqual(json_response["results"][0], expected_response)
+
+
+class TestCertificateAimAutocomplete(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.super_user = SuperUserFactory()
+        cls.url = reverse("certificate_aim_autocomplete")
+        cls.certificate_aim = CertificateAimFactory(
+            code=1234,
+            section=5,
+            description="description",
+        )
+
+    def test_user_not_logged(self):
+        self.client.logout()
+        response = self.client.get(self.url, data={'q': '1234'})
+        json_response = str(response.content, encoding='utf8')
+        results = json.loads(json_response)['results']
+        self.assertEqual(results, [])
+
+    def test_when_param_is_digit_assert_searching_on_code(self):
+        # When searching on "code"
+        self.client.force_login(user=self.super_user)
+        response = self.client.get(self.url, data={'q': '1234'})
+        self._assert_result_is_correct(response)
+
+    def test_assert_searching_on_description(self):
+        # When searching on "description"
+        self.client.force_login(user=self.super_user)
+        response = self.client.get(self.url, data={'q': 'descr'})
+        self._assert_result_is_correct(response)
+
+    def test_with_filter_by_section(self):
+        self.client.force_login(user=self.super_user)
+        response = self.client.get(self.url, data={'forward': '{"section": "5"}'})
+        self._assert_result_is_correct(response)
+
+    def _assert_result_is_correct(self, response):
+        self.assertEqual(response.status_code, 200)
+        json_response = str(response.content, encoding='utf8')
+        results = json.loads(json_response)['results']
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0]['id'], self.certificate_aim.code)
